requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.2
joblib>=1.3
xgboost>=1.7
python-snap7>=1.0
PyYAML>=6.0
fastapi>=0.95
uvicorn>=0.20
httpx>=0.24
